FuncApp/requirements.txt
# Azure SDKs
azure-functions
azure-identity
azure-mgmt-storage
azure-mgmt-resource
